backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pandas
openpyxl
google-genai
python-dotenv
